sdk: split query_verbs into module directory (7/9: extract sessions)
Move session_cost, inferences, sessions_list, and helpers to
query_verbs/sessions.rs.
